feat(gsdk): introduce module api
