+++ gsdk/src/rpc.rs
@@ -18,7 +18,7 @@
 
 //! gear api rpc methods
 #![allow(clippy::too_many_arguments)]
-use crate::{result::Result, types::GasInfo, Api};
+use crate::{result::Result, Api, GasInfo};
 use gear_core::ids::{CodeId, MessageId, ProgramId};
 use sp_core::H256;
 use subxt::rpc::rpc_params;
